chore(deps): split `optional-cmd-common` pkg into 2 packages - rename optional-cmd-common to `filter-packages` & move Profiler into its own `profiler` package
