+++ package.json
@@ -59,7 +59,8 @@
     "@lerna-lite/info": "workspace:*",
     "@lerna-lite/list": "workspace:*",
     "@lerna-lite/listable": "workspace:*",
-    "@lerna-lite/optional-cmd-common": "workspace:*",
+    "@lerna-lite/filter-packages": "workspace:*",
+    "@lerna-lite/profiler": "workspace:*",
     "@lerna-lite/publish": "workspace:*",
     "@lerna-lite/run": "workspace:*",
     "@lerna-lite/version": "workspace:*",
