+++ packages/filter-packages/README.md
@@ -1,16 +1,17 @@
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
-[![npm](https://img.shields.io/npm/dy/@lerna-lite/optional-cmd-common?color=forest)](https://www.npmjs.com/package/@lerna-lite/optional-cmd-common)
-[![npm](https://img.shields.io/npm/v/@lerna-lite/optional-cmd-common.svg?logo=npm&logoColor=fff&label=npm)](https://www.npmjs.com/package/@lerna-lite/optional-cmd-common)
+[![npm](https://img.shields.io/npm/dy/@lerna-lite/filter-packages?color=forest)](https://www.npmjs.com/package/@lerna-lite/filter-packages)
+[![npm](https://img.shields.io/npm/v/@lerna-lite/filter-packages.svg?logo=npm&logoColor=fff&label=npm)](https://www.npmjs.com/package/@lerna-lite/filter-packages)
 
-## Lerna-Lite common utils for optional commands (Exec/List/Run)
+## Lerna-Lite filtering utils for optional commands (Exec/List/Run/Watch)
 
-#### @lerna-lite/optional-cmd-common
+#### @lerna-lite/filter-packages
 
 Shared common code for the following optional commands:
 
 - [@lerna-lite/exec](https://github.com/lerna-lite/lerna-lite/tree/main/packages/exec)
 - [@lerna-lite/list](https://github.com/lerna-lite/lerna-lite/tree/main/packages/list)
 - [@lerna-lite/run](https://github.com/lerna-lite/lerna-lite/tree/main/packages/run)
+- [@lerna-lite/watch](https://github.com/lerna-lite/lerna-lite/tree/main/packages/watch)
 
 ### Installation
