+++ packages/filter-packages/src/__fixtures__/filtering/lerna.json
