+++ packages/filter-packages/src/__fixtures__/filtering/packages/package-1/package.json
