+++ packages/filter-packages/src/__fixtures__/filtering/packages/package-2/package.json
