+++ packages/filter-packages/src/__fixtures__/filtering/packages/package-3/package.json
