+++ packages/filter-packages/src/__fixtures__/filtering/packages/package-4/package.json
