+++ packages/filter-packages/src/__fixtures__/filtering/packages/package-5/package.json
