+++ packages/filter-packages/src/__tests__/index.spec.ts
@@ -1,3 +1,4 @@
+import 'jest-extended';
 import { ExecOpts } from '@lerna-lite/core';
 import * as entry from '../index';
 
@@ -7,7 +8,6 @@ describe('Testing common lib entry point', () => {
   });
 
   it('should have all exported object defined', () => {
-    expect(typeof entry.Profiler).toBe('function');
     expect(typeof entry.filterPackages).toBe('function');
     expect(typeof entry.getFilteredPackages).toBe('function');
   });
