+++ packages/filter-packages/src/index.ts
@@ -1,3 +1,3 @@
 export * from './filter-packages';
 export * from './get-filtered-packages';
-export * from './profiler';
+export * from './models';
