+++ packages/filter-packages/src/models/index.ts
@@ -13,18 +13,3 @@ export interface FilterOptions {
   includeDependencies: boolean;
   includeMergedTags: boolean;
 }
-
-export interface ProfilerConfig {
-  concurrency: number;
-  log?: typeof npmlog;
-  outputDirectory?: string;
-}
-
-export interface TraceEvent {
-  name: string;
-  ph: string;
-  ts: number;
-  pid: number;
-  tid: number;
-  dur: number;
-}
