+++ packages/profiler/README.md
@@ -1,0 +1,17 @@
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+[![npm](https://img.shields.io/npm/dy/@lerna-lite/profiler?color=forest)](https://www.npmjs.com/package/@lerna-lite/profiler)
+[![npm](https://img.shields.io/npm/v/@lerna-lite/profiler.svg?logo=npm&logoColor=fff&label=npm)](https://www.npmjs.com/package/@lerna-lite/profiler)
+
+## Lerna-Lite common utils for optional commands (Exec/List/Run)
+
+#### @lerna-lite/profiler
+
+Shared common code for the following optional commands:
+
+- [@lerna-lite/exec](https://github.com/lerna-lite/lerna-lite/tree/main/packages/exec)
+- [@lerna-lite/list](https://github.com/lerna-lite/lerna-lite/tree/main/packages/list)
+- [@lerna-lite/run](https://github.com/lerna-lite/lerna-lite/tree/main/packages/run)
+
+### Installation
+
+Follow the instruction provided in the main [README](https://github.com/lerna-lite/lerna-lite#installation) page.
