+++ packages/run/package.json
@@ -33,7 +33,8 @@
   },
   "dependencies": {
     "@lerna-lite/core": "workspace:*",
-    "@lerna-lite/optional-cmd-common": "workspace:*",
+    "@lerna-lite/filter-packages": "workspace:*",
+    "@lerna-lite/profiler": "workspace:*",
     "chalk": "^4.1.2",
     "fs-extra": "^11.1.0",
     "npmlog": "^7.0.1",
