+++ packages/run/src/run-command.ts
@@ -8,12 +8,8 @@ import {
   runTopologically,
   ValidationError,
 } from '@lerna-lite/core';
-import {
-  FilterOptions,
-  generateProfileOutputPath,
-  getFilteredPackages,
-  Profiler,
-} from '@lerna-lite/optional-cmd-common';
+import { FilterOptions, getFilteredPackages } from '@lerna-lite/filter-packages';
+import { generateProfileOutputPath, Profiler } from '@lerna-lite/profiler';
 import chalk from 'chalk';
 import { existsSync } from 'fs-extra';
 import pMap from 'p-map';
