+++ packages/watch/src/__tests__/watch-command.spec.ts
@@ -9,11 +9,6 @@ jest.mock('@lerna-lite/core', () => ({
   QueryGraph: jest.requireActual('../../../core/src/utils/query-graph').QueryGraph,
 }));
 
-jest.mock('@lerna-lite/optional-cmd-common', () => ({
-  ...(jest.requireActual('@lerna-lite/optional-cmd-common') as any),
-  ...jest.requireActual('../../../optional-cmd-common/src/lib/profiler'), // test with the real Profiler for test coverage as well
-}));
-
 let watchAddHandler;
 let watchAddDirHandler;
 let watchUnlinkHandler;
